refactor: get rid of validation-function-params crate
Everything about the VFPs has been moved into a module of runtime
