--- validation-function-params/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "cumulus-validation-function-params"
-version = "2.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-license = "GPL-3.0"
-homepage = "https://substrate.dev"
-repository = "https://github.com/paritytech/cumulus/"
-description = "Inherent for validation function parameters"
-
-[dependencies]
-sp-inherents = { git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
-codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false, features = ["derive"] }
-cumulus-runtime = { path = "../runtime", default-features = false }
-
-
-[features]
-default = [ "std" ]
-std = [
-	"codec/std",
-	"sp-inherents/std",
-	"sp-runtime/std",
-]
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
